Reafactor - 필요 없는 파일 삭제, 파일 패키지 수정
